Migrate to MIT license
Before the project code was licensed under the Apache 2.0 [1] license.
This often caused problems when developers wanted to use the project or
code parts of it in another project (mainly code editors and UI
frameworks) licensed under a less restricted license.

Therefore the project migrated to the MIT license [2] which is more
open, unrestricted and the most used license for open source projects
(Facebook recently also re-licensed React [3]).

[1]: http://www.apache.org/licenses/LICENSE-2.0
[2]: https://choosealicense.com/licenses/mit
[3]: https://code.facebook.com/posts/300798627056246/relicensing-react-jest-flow-and-immutable-js

GH-7